theta to radian's type : double -> float

+++ FdF/calc_rotate.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   calc_rotate.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seongtki <seongtki@student.42seoul.>       +#+  +:+       +#+        */
+/*   By: seongtki <seongtki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/07 18:04:54 by seongtki          #+#    #+#             */
-/*   Updated: 2022/09/07 18:05:10 by seongtki         ###   ########.fr       */
+/*   Updated: 2022/12/29 21:33:33 by seongtki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-void	rotate_x(int *y, int *z, double radian)
+void	rotate_x(int *y, int *z, float radian)
 {
 	int	prev_y;
 
@@ -21,7 +21,7 @@ void	rotate_x(int *y, int *z, double radian)
 	*z = -prev_y * sin(radian) + (*z) * cos(radian);
 }
 
-void	rotate_y(int *x, int *z, double radian)
+void	rotate_y(int *x, int *z, float radian)
 {
 	int	prev_x;
 
@@ -30,7 +30,7 @@ void	rotate_y(int *x, int *z, double radian)
 	*z = -prev_x * sin(radian) + (*z) * cos(radian);
 }
 
-void	rotate_z(int *x, int *y, double radian)
+void	rotate_z(int *x, int *y, float radian)
 {
 	int	prev_x;
 	int	prev_y;
